Split up Postprocessor and GenericDataOut
