+++ include/meltpooldg/advection_diffusion/advection_diffusion_adaflo_wrapper.hpp
@@ -12,6 +12,7 @@
 #  include <meltpooldg/advection_diffusion/advection_diffusion_operation_base.hpp>
 #  include <meltpooldg/interface/scratch_data.hpp>
 #  include <meltpooldg/interface/simulationbase.hpp>
+#  include <meltpooldg/utilities/generic_data_out.hpp>
 
 #  include <adaflo/diagonal_preconditioner.h>
 #  include <adaflo/level_set_okz_advance_concentration.h>
