+++ include/meltpooldg/advection_diffusion/advection_diffusion_operation_base.hpp
@@ -10,6 +10,7 @@
 
 #include <meltpooldg/interface/parameters.hpp>
 #include <meltpooldg/interface/scratch_data.hpp>
+#include <meltpooldg/utilities/generic_data_out.hpp>
 
 namespace MeltPoolDG
 {
