+++ include/meltpooldg/flow/adaflo_wrapper.hpp
@@ -15,6 +15,7 @@
 #  include <meltpooldg/flow/flow_base.hpp>
 #  include <meltpooldg/interface/scratch_data.hpp>
 #  include <meltpooldg/interface/simulationbase.hpp>
+#  include <meltpooldg/utilities/generic_data_out.hpp>
 #  include <meltpooldg/utilities/utilityfunctions.hpp>
 #  include <meltpooldg/utilities/vector_tools.hpp>
