+++ include/meltpooldg/heat/heat_transfer_operator.hpp
@@ -12,6 +12,7 @@
 #include <meltpooldg/interface/boundaryconditions.hpp>
 #include <meltpooldg/interface/operator_base.hpp>
 #include <meltpooldg/interface/parameters.hpp>
+#include <meltpooldg/utilities/generic_data_out.hpp>
 #include <meltpooldg/utilities/physical_constants.hpp>
 #include <meltpooldg/utilities/vector_tools.hpp>
