+++ include/meltpooldg/level_set/level_set_operation.hpp
@@ -11,6 +11,7 @@
 #include <meltpooldg/curvature/curvature_operation_base.hpp>
 #include <meltpooldg/level_set/level_set_operator_with_phase_change.hpp>
 #include <meltpooldg/reinitialization/reinitialization_operation_base.hpp>
+#include <meltpooldg/utilities/generic_data_out.hpp>
 #include <meltpooldg/utilities/timeiterator.hpp>
 
 namespace MeltPoolDG::LevelSet
