+++ include/meltpooldg/level_set/level_set_problem.hpp
@@ -10,6 +10,7 @@
 #include <meltpooldg/interface/scratch_data.hpp>
 #include <meltpooldg/interface/simulationbase.hpp>
 #include <meltpooldg/level_set/level_set_operation.hpp>
+#include <meltpooldg/utilities/postprocessor.hpp>
 #include <meltpooldg/utilities/timeiterator.hpp>
 
 namespace MeltPoolDG::LevelSet
