+++ include/meltpooldg/reinitialization/reinitialization_operation_adaflo_wrapper.hpp
@@ -14,6 +14,7 @@
 #  include <meltpooldg/normal_vector/normal_vector_operation_adaflo_wrapper.hpp>
 #  include <meltpooldg/reinitialization/reinitialization_operation_base.hpp>
 #  include <meltpooldg/utilities/conditional_ostream.hpp>
+#  include <meltpooldg/utilities/generic_data_out.hpp>
 #  include <meltpooldg/utilities/vector_tools.hpp>
 
 #  include <adaflo/diagonal_preconditioner.h>
