+++ include/meltpooldg/utilities/postprocessor.hpp
@@ -16,6 +16,7 @@
 
 #include <meltpooldg/interface/parameters.hpp>
 #include <meltpooldg/utilities/conditional_ostream.hpp>
+#include <meltpooldg/utilities/generic_data_out.hpp>
 #include <meltpooldg/utilities/utilityfunctions.hpp>
 
 #include <filesystem>
@@ -26,34 +27,6 @@ using namespace dealii;
 
 namespace MeltPoolDG
 {
-  template <int dim>
-  class GenericDataOut
-  {
-  public:
-    using VectorType = LinearAlgebra::distributed::Vector<double>;
-
-    GenericDataOut() = default;
-
-    void
-    add_data_vector(const DoFHandler<dim> &         dof_handler,
-                    const VectorType &              data,
-                    const std::vector<std::string> &names,
-                    const std::vector<DataComponentInterpretation::DataComponentInterpretation>
-                      &data_component_interpretation =
-                        std::vector<DataComponentInterpretation::DataComponentInterpretation>());
-
-    void
-    add_data_vector(const DoFHandler<dim> &dof_handler,
-                    const VectorType &     data,
-                    const std::string &    name);
-
-    std::vector<std::tuple<const DoFHandler<dim> *,
-                           const VectorType *,
-                           const std::vector<std::string>,
-                           std::vector<DataComponentInterpretation::DataComponentInterpretation>>>
-      entries;
-  };
-
   template <int dim>
   class Postprocessor
   {
